App Routerへの移行を完了
- Pages RouterからApp Routerへ完全移行
- src/pages/ディレクトリをsrc/app/に移行
- _app.tsxをlayout.tsxに変換
- API RoutesをRoute Handlers形式に変換
  - /api/health → app/api/health/route.ts
  - /api/projects → app/api/projects/route.ts
  - /api/projects/[id] → app/api/projects/[id]/route.ts
- テストファイルをApp Router対応に更新
- .js拡張子のインポートを削除（Next.jsビルド対応）
- Server Actions利用可能な環境を整備
