+++ src/database/index.ts
@@ -2,8 +2,8 @@
 // 全てのデータベース機能をこのファイルから提供
 
 // 接続とコア機能
-export { closeConnection, db, pool, testConnection } from "./connection.js";
+export { closeConnection, db, pool, testConnection } from "./connection";
 // リポジトリクラスとインスタンス
-export * from "./repositories/index.js";
+export * from "./repositories/index";
 // スキーマ定義
-export * from "./schema/index.js";
+export * from "./schema/index";
