+++ src/database/schema/commits.ts
@@ -8,7 +8,7 @@ import {
 	uniqueIndex,
 	varchar,
 } from "drizzle-orm/pg-core";
-import { projects } from "./projects.js";
+import { projects } from "./projects";
 
 /**
  * GitLabコミット情報テーブル
