+++ src/database/schema/index.ts
@@ -1,5 +1,5 @@
 // データベーススキーマ定義
 
-export * from "./commits.js";
-export * from "./projects.js";
-export * from "./sync-logs.js";
+export * from "./commits";
+export * from "./projects";
+export * from "./sync-logs";
