+++ src/database/schema/sync-logs.ts
@@ -7,7 +7,7 @@ import {
 	text,
 	timestamp,
 } from "drizzle-orm/pg-core";
-import { projects } from "./projects.js";
+import { projects } from "./projects";
 
 /**
  * 同期タイプ定義
